extracted classes to files; moved hooks to dedicated namespaces
