+++ EventHook/Hooks/ClipBoard/ClipBoardHook.cs
