+++ EventHook/Hooks/Keyboard/InterceptKeys.cs
@@ -1,0 +1,220 @@
+using System;
+using System.Diagnostics;
+using System.Runtime.InteropServices;
+using System.Text;
+
+namespace EventHook.Hooks.Keyboard
+{
+    /// <summary>
+    ///     Winapi Key interception helper class.
+    /// </summary>
+    internal static class InterceptKeys
+    {
+        internal delegate IntPtr LowLevelKeyboardProc(int nCode, UIntPtr wParam, IntPtr lParam);
+
+        /// <summary>
+        ///     Key event
+        /// </summary>
+        internal enum KeyEvent
+        {
+            /// <summary>
+            ///     Key down
+            /// </summary>
+            WM_KEYDOWN = 256,
+
+            /// <summary>
+            ///     Key up
+            /// </summary>
+            WM_KEYUP = 257,
+
+            /// <summary>
+            ///     System key up
+            /// </summary>
+            WM_SYSKEYUP = 261,
+
+            /// <summary>
+            ///     System key down
+            /// </summary>
+            WM_SYSKEYDOWN = 260
+        }
+
+        internal static int WH_KEYBOARD_LL = 13;
+
+        internal static IntPtr SetHook(LowLevelKeyboardProc proc)
+        {
+            using (var curProcess = Process.GetCurrentProcess())
+            using (var curModule = curProcess.MainModule)
+            {
+                return SetWindowsHookEx(WH_KEYBOARD_LL, proc, GetModuleHandle(curModule.ModuleName), 0);
+            }
+        }
+
+        [DllImport("user32.dll", CharSet = CharSet.Auto, SetLastError = true)]
+        internal static extern IntPtr SetWindowsHookEx(int idHook, LowLevelKeyboardProc lpfn, IntPtr hMod, uint dwThreadId);
+
+        [DllImport("user32.dll", CharSet = CharSet.Auto, SetLastError = true)]
+        [return: MarshalAs(UnmanagedType.Bool)]
+        internal static extern bool UnhookWindowsHookEx(IntPtr hhk);
+
+        [DllImport("user32.dll", CharSet = CharSet.Auto, SetLastError = true)]
+        internal static extern IntPtr CallNextHookEx(IntPtr hhk, int nCode, UIntPtr wParam, IntPtr lParam);
+
+        [DllImport("kernel32.dll", CharSet = CharSet.Auto, SetLastError = true)]
+        internal static extern IntPtr GetModuleHandle(string lpModuleName);
+
+        #region Convert VKCode to string
+
+        // Note: Sometimes single VKCode represents multiple chars, thus string. 
+        // E.g. typing "^1" (notice that when pressing 1 the both characters appear, 
+        // because of this behavior, "^" is called dead key)
+
+        [DllImport("user32.dll")]
+        private static extern int ToUnicodeEx(uint wVirtKey, uint wScanCode, byte[] lpKeyState,
+            [Out, MarshalAs(UnmanagedType.LPWStr)] StringBuilder pwszBuff, int cchBuff, uint wFlags, IntPtr dwhkl);
+
+        [DllImport("user32.dll")]
+        private static extern bool GetKeyboardState(byte[] lpKeyState);
+
+        [DllImport("user32.dll")]
+        private static extern uint MapVirtualKeyEx(uint uCode, uint uMapType, IntPtr dwhkl);
+
+        [DllImport("user32.dll", CharSet = CharSet.Auto, ExactSpelling = true)]
+        private static extern IntPtr GetKeyboardLayout(uint dwLayout);
+
+        [DllImport("User32.dll")]
+        private static extern IntPtr GetForegroundWindow();
+
+        [DllImport("User32.dll")]
+        private static extern uint GetWindowThreadProcessId(IntPtr hWnd, out uint lpdwProcessId);
+
+        [DllImport("user32.dll")]
+        private static extern bool AttachThreadInput(uint idAttach, uint idAttachTo, bool fAttach);
+
+        [DllImport("kernel32.dll")]
+        private static extern uint GetCurrentThreadId();
+
+        private static uint _lastVkCode;
+        private static uint _lastScanCode;
+        private static byte[] _lastKeyState = new byte[255];
+        private static bool _lastIsDead;
+
+        /// <summary>
+        ///     Convert VKCode to Unicode.
+        ///     <remarks>isKeyDown is required for because of keyboard state inconsistencies!</remarks>
+        /// </summary>
+        /// <param name="vkCode">VKCode</param>
+        /// <param name="isKeyDown">Is the key down event?</param>
+        /// <returns>String representing single unicode character.</returns>
+        internal static string VkCodeToString(uint vkCode, bool isKeyDown)
+        {
+            // ToUnicodeEx needs StringBuilder, it populates that during execution.
+            var sbString = new StringBuilder(5);
+
+            var bKeyState = new byte[255];
+            bool bKeyStateStatus;
+            var isDead = false;
+
+            // Gets the current windows window handle, threadID, processID
+            var currentHWnd = GetForegroundWindow();
+            uint currentProcessId;
+            var currentWindowThreadId = GetWindowThreadProcessId(currentHWnd, out currentProcessId);
+
+            // This programs Thread ID
+            var thisProgramThreadId = GetCurrentThreadId();
+
+            // Attach to active thread so we can get that keyboard state
+            if (AttachThreadInput(thisProgramThreadId, currentWindowThreadId, true))
+            {
+                // Current state of the modifiers in keyboard
+                bKeyStateStatus = GetKeyboardState(bKeyState);
+
+                // Detach
+                AttachThreadInput(thisProgramThreadId, currentWindowThreadId, false);
+            }
+            else
+            {
+                // Could not attach, perhaps it is this process?
+                bKeyStateStatus = GetKeyboardState(bKeyState);
+            }
+
+            // On failure we return empty string.
+            if (!bKeyStateStatus)
+                return "";
+
+            // Gets the layout of keyboard
+            var hkl = GetKeyboardLayout(currentWindowThreadId);
+
+            // Maps the virtual keycode
+            var lScanCode = MapVirtualKeyEx(vkCode, 0, hkl);
+
+            // Keyboard state goes inconsistent if this is not in place. In other words, we need to call above commands in UP events also.
+            if (!isKeyDown)
+                return "";
+
+            // Converts the VKCode to unicode
+            var relevantKeyCountInBuffer = ToUnicodeEx(vkCode, lScanCode, bKeyState, sbString, sbString.Capacity, 0, hkl);
+
+            var ret = string.Empty;
+
+            switch (relevantKeyCountInBuffer)
+            {
+                // Dead keys (^,`...)
+                case -1:
+                    isDead = true;
+
+                    // We must clear the buffer because ToUnicodeEx messed it up, see below.
+                    ClearKeyboardBuffer(vkCode, lScanCode, hkl);
+                    break;
+
+                case 0:
+                    break;
+
+                // Single character in buffer
+                case 1:
+                    ret = sbString[0].ToString();
+                    break;
+
+                // Two or more (only two of them is relevant)
+                default:
+                    ret = sbString.ToString().Substring(0, 2);
+                    break;
+            }
+
+            // We inject the last dead key back, since ToUnicodeEx removed it.
+            // More about this peculiar behavior see e.g: 
+            //   http://www.experts-exchange.com/Programming/System/Windows__Programming/Q_23453780.html
+            //   http://blogs.msdn.com/michkap/archive/2005/01/19/355870.aspx
+            //   http://blogs.msdn.com/michkap/archive/2007/10/27/5717859.aspx
+            if (_lastVkCode != 0 && _lastIsDead)
+            {
+                var sbTemp = new StringBuilder(5);
+                ToUnicodeEx(_lastVkCode, _lastScanCode, _lastKeyState, sbTemp, sbTemp.Capacity, 0, hkl);
+                _lastVkCode = 0;
+
+                return ret;
+            }
+
+            // Save these
+            _lastScanCode = lScanCode;
+            _lastVkCode = vkCode;
+            _lastIsDead = isDead;
+            _lastKeyState = (byte[]) bKeyState.Clone();
+
+            return ret;
+        }
+
+        private static void ClearKeyboardBuffer(uint vk, uint sc, IntPtr hkl)
+        {
+            var sb = new StringBuilder(10);
+
+            int rc;
+            do
+            {
+                var lpKeyStateNull = new byte[255];
+                rc = ToUnicodeEx(vk, sc, lpKeyStateNull, sb, sb.Capacity, 0, hkl);
+            } while (rc < 0);
+        }
+
+        #endregion
+    }
+}
