+++ EventHook/Hooks/Mouse/RawMouseEventArgs.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using EventHook.Hooks.Library;
+
+namespace EventHook.Hooks.Mouse
+{
+    internal class RawMouseEventArgs : EventArgs
+    {
+        internal Msgs Message { get; set; }
+        internal POINT Point { get; set; }
+    }
+}
