+++ EventHook/Hooks/PrintQueue/PrintQueueHook.cs
@@ -7,48 +7,9 @@ using System.Threading;
 using EventHook.Hooks.Library;
 using Microsoft.Win32.SafeHandles;
 
-namespace EventHook.Hooks
+namespace EventHook.Hooks.PrintQueue
 {
     //http://www.codeproject.com/Articles/51085/Monitor-jobs-in-a-printer-queue-NET
-    internal class PrintJobChangeEventArgs : EventArgs
-    {
-        internal PrintJobChangeEventArgs(int intJobID, string strJobName, JOBSTATUS jStatus, PrintSystemJobInfo objJobInfo)
-        {
-            _jobId = intJobID;
-            _jobName = strJobName;
-            _jobStatus = jStatus;
-            _jobInfo = objJobInfo;
-        }
-
-        internal int JobId
-        {
-            get { return _jobId; }
-        }
-
-        internal string JobName
-        {
-            get { return _jobName; }
-        }
-
-        internal JOBSTATUS JobStatus
-        {
-            get { return _jobStatus; }
-        }
-
-        internal PrintSystemJobInfo JobInfo
-        {
-            get { return _jobInfo; }
-        }
-
-        #region private variables
-
-        private readonly int _jobId;
-        private readonly string _jobName;
-        private readonly JOBSTATUS _jobStatus;
-        private readonly PrintSystemJobInfo _jobInfo;
-
-        #endregion
-    }
 
     internal delegate void PrintJobStatusChanged(object sender, PrintJobChangeEventArgs e);
 
@@ -89,11 +50,11 @@ namespace EventHook.Hooks
 
         internal void Start()
         {
-            OpenPrinter(SpoolerName, out _printerHandle, 0);
+            WinSpool.OpenPrinter(SpoolerName, out _printerHandle, 0);
             if (_printerHandle != IntPtr.Zero)
             {
                 //We got a valid Printer handle.  Let us register for change notification....
-                _changeHandle = FindFirstPrinterChangeNotification(_printerHandle,
+                _changeHandle = WinSpool.FindFirstPrinterChangeNotification(_printerHandle,
                     (int) PRINTER_CHANGES.PRINTER_CHANGE_JOB, 0, _notifyOptions);
                 // We have successfully registered for change notification.  Let us capture the handle...
                 _mrEvent.SafeWaitHandle = new SafeWaitHandle(_changeHandle, true);
@@ -103,7 +64,7 @@ namespace EventHook.Hooks
                     true);
             }
 
-            _spooler = new PrintQueue(new PrintServer(), SpoolerName);
+            _spooler = new System.Printing.PrintQueue(new PrintServer(), SpoolerName);
             foreach (var psi in _spooler.GetPrintJobInfoCollection())
             {
                 _objJobDict[psi.JobIdentifier] = psi.Name;
@@ -118,7 +79,7 @@ namespace EventHook.Hooks
         {
             if (_printerHandle != IntPtr.Zero)
             {
-                ClosePrinter((int) _printerHandle);
+                WinSpool.ClosePrinter((int) _printerHandle);
                 _printerHandle = IntPtr.Zero;
             }
         }
@@ -136,7 +97,7 @@ namespace EventHook.Hooks
             _notifyOptions.Count = 1;
             var pdwChange = 0;
             IntPtr pNotifyInfo;
-            var bResult = FindNextPrinterChangeNotification(_changeHandle, out pdwChange, _notifyOptions,
+            var bResult = WinSpool.FindNextPrinterChangeNotification(_changeHandle, out pdwChange, _notifyOptions,
                 out pNotifyInfo);
             //If the Printer Change Notification Call did not give data, exit code
             if ((bResult == false) || (((int) pNotifyInfo) == 0)) return;
@@ -183,7 +144,7 @@ namespace EventHook.Hooks
                     PrintSystemJobInfo pji = null;
                     try
                     {
-                        _spooler = new PrintQueue(new PrintServer(), SpoolerName);
+                        _spooler = new System.Printing.PrintQueue(new PrintServer(), SpoolerName);
                         pji = _spooler.GetJob(intJobId);
                         if (!_objJobDict.ContainsKey(intJobId))
                             _objJobDict[intJobId] = pji.Name;
@@ -216,47 +177,6 @@ namespace EventHook.Hooks
         }
 
         #endregion
-
-        #region DLL Import Functions
-
-        [DllImport("winspool.drv", EntryPoint = "OpenPrinterA", SetLastError = true, CharSet = CharSet.Ansi,
-            ExactSpelling = true, CallingConvention = CallingConvention.StdCall)]
-        internal static extern bool OpenPrinter(string pPrinterName,
-            out IntPtr phPrinter,
-            int pDefault);
-
-
-        [DllImport("winspool.drv", EntryPoint = "ClosePrinter",
-            SetLastError = true,
-            ExactSpelling = true,
-            CallingConvention = CallingConvention.StdCall)]
-        internal static extern bool ClosePrinter
-            (int hPrinter);
-
-        [DllImport("winspool.drv",
-            EntryPoint = "FindFirstPrinterChangeNotification",
-            SetLastError = true, CharSet = CharSet.Ansi,
-            ExactSpelling = true,
-            CallingConvention = CallingConvention.StdCall)]
-        internal static extern IntPtr FindFirstPrinterChangeNotification
-            ([In] IntPtr hPrinter,
-                [In] int fwFlags,
-                [In] int fwOptions,
-                [In, MarshalAs(UnmanagedType.LPStruct)] PRINTER_NOTIFY_OPTIONS pPrinterNotifyOptions);
-
-        [DllImport("winspool.drv", EntryPoint = "FindNextPrinterChangeNotification",
-            SetLastError = true, CharSet = CharSet.Ansi,
-            ExactSpelling = false,
-            CallingConvention = CallingConvention.StdCall)]
-        internal static extern bool FindNextPrinterChangeNotification
-            ([In] IntPtr hChangeObject,
-                [Out] out int pdwChange,
-                [In, MarshalAs(UnmanagedType.LPStruct)] PRINTER_NOTIFY_OPTIONS pPrinterNotifyOptions,
-                [Out] out IntPtr lppPrinterNotifyInfo
-            );
-
-        #endregion
-
         #region private variables
 
         private IntPtr _printerHandle = IntPtr.Zero;
@@ -266,7 +186,7 @@ namespace EventHook.Hooks
         private IntPtr _changeHandle = IntPtr.Zero;
         private readonly PRINTER_NOTIFY_OPTIONS _notifyOptions = new PRINTER_NOTIFY_OPTIONS();
         private readonly Dictionary<int, string> _objJobDict = new Dictionary<int, string>();
-        private PrintQueue _spooler;
+        private System.Printing.PrintQueue _spooler;
 
         #endregion
     }
